requirements.txt
# requirements.txt
langchain
langchain-openai
langchain-community
clickhouse-sqlalchemy
sqlalchemy
python-dotenv